chore: refactor types
- Normalise filenames
- Normalise type names
- Correct imports and exports to use the index
- Move types used across site to `ui.ts`
