+++ src/components/home/BlogCard.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import { BlogPost } from "@/types/BlogItem";
+import { BlogPost } from "@/types";
 
 /**
  * A card component for displaying a single blog post preview.
