+++ src/types/index.ts
@@ -1,2 +1,4 @@
-export * from "./Image";
-export * from "./ProjectMeta";
+export * from "./about";
+export * from "./blog";
+export * from "./project";
+export * from "./ui";
